+++ Server/Controllers/GameHub.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.SignalR;
 using FreeTabletop.Server.Models;
@@ -71,8 +70,8 @@ namespace FreeTabletop.Server.Controllers
             }
         }
 
-        [HubMethodName("Room:EnableCell")]
-        public async Task EnableCell(int cellIndex)
+        [HubMethodName("Room:ChangeCellStyle")]
+        public void ChangeCellStyles(int index, string style)
         {
             Player player = GetPlayer(Context.ConnectionId);
             if (player != null && player.IsGameMaster)
@@ -80,8 +79,8 @@ namespace FreeTabletop.Server.Controllers
                 Room room = GetRoom(player.RoomCode);
                 if (room != null)
                 {
-                    room.EnableCell(cellIndex);
-                    await Clients.Group(room.RoomCode).SendAsync("Tabletop:UpdateCellVisiblity", cellIndex);
+                    room.UpdateCellStyle(index, style);
+                    Clients.Group(room.RoomCode).SendAsync("Tabletop:SyncCells", index, style);
                 }
             }
         }